added simple contextvars to task and api wrappers
