+++ app/api/middleware.py
@@ -3,13 +3,13 @@
 import uuid
 from collections.abc import Awaitable, Callable
 
-import structlog
 from fastapi import status
 from starlette.applications import Starlette
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.requests import Request
 from starlette.responses import Response
 from structlog import get_logger
+from structlog.contextvars import bind_contextvars, unbind_contextvars
 
 
 class LoggerMiddleware(BaseHTTPMiddleware):
@@ -45,8 +45,7 @@ class LoggerMiddleware(BaseHTTPMiddleware):
             The response from the next middleware or route handler.
 
         """
-        structlog.contextvars.clear_contextvars()
-        structlog.contextvars.bind_contextvars(
+        bind_contextvars(
             path=request.url.path,
             method=request.method,
             client_host=request.client and request.client.host,
@@ -55,7 +54,7 @@ class LoggerMiddleware(BaseHTTPMiddleware):
 
         try:
             response = await call_next(request)
-            structlog.contextvars.bind_contextvars(status_code=response.status_code)
+            bind_contextvars(status_code=response.status_code)
 
             if (
                 status.HTTP_400_BAD_REQUEST
@@ -70,6 +69,12 @@ class LoggerMiddleware(BaseHTTPMiddleware):
 
         except Exception:
             self.logger.exception("Unhandled exception in request")
+            unbind_contextvars(
+                "path", "method", "client_host", "request_id", "status_code"
+            )
             raise
         else:
+            unbind_contextvars(
+                "path", "method", "client_host", "request_id", "status_code"
+            )
             return response
