+++ app/core/telemetry/fastapi.py
@@ -7,8 +7,15 @@ from opentelemetry import trace
 from starlette.applications import Starlette
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.routing import Match
+from structlog import get_logger
+from structlog.contextvars import (
+    bind_contextvars,
+    clear_contextvars,
+)
 
 from app.core.telemetry.attributes import Attributes
+
+logger = get_logger(__name__)
 
 
 class FastAPITracingMiddleware(BaseHTTPMiddleware):
@@ -38,29 +45,38 @@ class FastAPITracingMiddleware(BaseHTTPMiddleware):
             The response from the next middleware or route handler.
 
         """
-        # Get the current span
         current_span = trace.get_current_span()
 
-        if current_span.is_recording():
-            # Add query string if present
-            if request.url.query:
-                # Add individual query parameters as attributes using FastAPI
+        if request.url.query:
+            # Add individual query parameters as attributes using FastAPI
+            if current_span.is_recording():
                 for key, value in request.query_params.items():
                     current_span.set_attribute(
                         f"{Attributes.HTTP_REQUEST_QUERY_PARAMS}.{key}", value
                     )
+            bind_contextvars(**request.query_params)
 
-            # Can't access path parameters directly in middleware.
-            # This is what starlette does internally.
-            # https://github.com/fastapi/fastapi/issues/861
-            # https://github.com/encode/starlette/blob/5c43dde0ec0917673bb280bcd7ab0c37b78061b7/starlette/routing.py#L544
-            routes = request.app.router.routes
-            for route in routes:
-                match, scope = route.matches(request)
-                if match == Match.FULL:
+        # Can't access path parameters directly in middleware.
+        # This is what starlette does internally.
+        # https://github.com/fastapi/fastapi/issues/861
+        # https://github.com/encode/starlette/blob/5c43dde0ec0917673bb280bcd7ab0c37b78061b7/starlette/routing.py#L544
+        routes = request.app.router.routes
+        for route in routes:
+            match, scope = route.matches(request)
+            if match == Match.FULL:
+                if current_span.is_recording():
                     for key, value in scope["path_params"].items():
                         current_span.set_attribute(
                             f"{Attributes.HTTP_REQUEST_PATH_PARAMS}.{key}", str(value)
                         )
+                bind_contextvars(**scope["path_params"])
 
-        return await call_next(request)
+        try:
+            result = await call_next(request)
+        except Exception:
+            clear_contextvars()
+            raise
+        else:
+            # This is guaranteed to be the outermost middleware so we clear all context
+            clear_contextvars()
+            return result
