+++ app/main.py
@@ -8,7 +8,7 @@ from structlog import get_logger
 
 from app.api.root import register_api
 from app.core.config import get_settings
-from app.core.logger import logger_configurer
+from app.core.telemetry.logger import logger_configurer
 from app.core.telemetry.otel import configure_otel
 from app.persistence.es.client import es_manager
 from app.persistence.sql.session import db_manager
